todo list

+++ src/Components/List.jsx
@@ -3,12 +3,11 @@ import todo from "./todo.png";
 import TodoList from "./TodoList";
 import "./List.css";
 import Footer from "./Footer";
-import ShowCompleted from "./ShowCompleted";
 
 const List = () => {
   const [inputList, setInputList] = useState("");
   const [items, setItems] = useState([]);
- 
+
   let CompletedFalse = items.filter((value) => !value.Completed);
 
   const inputChange = (e) => {
@@ -30,27 +29,33 @@ const List = () => {
     }
     setInputList("");
   };
- 
 
   const deleteItem = (id) => {
-    // console.log('deleted');
+    // console.log('deleted' , id);
     setItems((oldItems) => {
       return oldItems.filter((arrEle, index) => {
-        // afterChecked();
         return index !== id;
       });
     });
   };
- 
+
   const checked = (id) => {
-    
     let temp = [...items];
+    // let temp = items.filter((value) => !value.Completed);
     temp[id].Completed = true;
+    setItems(temp);
+  };
+
+  const unchecked = (id) => {
+    let temp = [...items];
+    // let temp = items.filter((value) => !value.Completed);
+    temp[id].Completed = false;
     setItems(temp);
   };
 
   const clearAll = () => {
     setItems([]);
+    setInputList("");
   };
 
   const keyupFunction = () => {
@@ -64,7 +69,6 @@ const List = () => {
     document.getElementById("add").style.backgroundColor = "#9b969b";
   };
 
-  
   //  let log = items.map(todo => todo.Completed)
   //  console.log("p render");
   //  console.log("length", items.length);
@@ -93,30 +97,24 @@ const List = () => {
       {CompletedFalse.length > 0 ? (
         <div>
           <div className="pandingItem">
-            <p>You have {items.length} panding item</p>
+            <p>You have {CompletedFalse.length} pending item</p>
           </div>
           <div className="middle">
-        
             <ul className="todoList">
-              {CompletedFalse.map((obj, index) => {
-
-                return (
-                  <TodoList
-                    text={obj.List}
-                    key={Math.random()}
-                    id={index}
-                    onSelect={deleteItem}
-                    checked={checked}
-                  />
-                );
+              {items.map((obj, index) => {
+                if (obj.Completed === false)
+                  return (
+                    <TodoList
+                      text={obj.List}
+                      key={Math.random()}
+                      id={index}
+                      onSelect={deleteItem}
+                      checked={checked}
+                    />
+                  );
               })}
             </ul>
-
-            
-
-            <Footer allItem={clearAll}  items1={items} setItems={setItems}/>
           </div>
-
         </div>
       ) : (
         <div className="todo">
@@ -124,8 +122,16 @@ const List = () => {
           <p className="todopera">Time to chill You have no todos.</p>
         </div>
       )}
-      
-      
+      {items.length > 0 && (
+        <Footer
+          allItem={clearAll}
+          items1={items}
+          setItems={setItems}
+          id={items.index}
+          deleteItem={deleteItem}
+          unchecked={unchecked}
+        />
+      )}
     </>
   );
 };
